separate implementation from test
