+++ pi_spigot.vcxproj
@@ -173,6 +173,9 @@
   <ItemGroup>
     <None Include="codecov.yml" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="pi_spigot\pi_spigot.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
